filter: mutable since
We extend filter to allow mutation of the since field. Since filters
are not normally modifiable in the sense of being able to add additional
fields, we create some machinery to mutate the since field if it exists,
otherwise we create a new filter with it.

We also greatly simplify many of the FilterBuilder constructors to use
iterators. Iterator all the things!

Signed-off-by: William Casarin <jb55@jb55.com>