filter: add custom note filtering
Add the ability to add custom filter conditions. These conditions are
executed for each matching note when nostrdb walks the index during
a query. If the callback returns false, the note is not added to
the query result.

This is useful for things like filtering note replies, etc

Changelog-Added: Added the ability to add custom filter conditions to Filter
Signed-off-by: William Casarin <jb55@jb55.com>